BUG#15930031 - SETTING SLOW_QUERY_LOG_FILE TO A VALUE WITHOUT                A "=" SIGN ENABLES GENERAL LOG
DESCRIPTION:

While checking server startup options from configuration file,
MySQL checks if the option is a struct option, by finding dot
('.') and checking if dot lies before or after '='. If dot is
before '=' then the startup option will be just after dot. If
dot lies after '=' then it is a part of argument for startup
option. If '=' is not in the startup option and dot is there,
then check_struct_option() should return pointer to the begin
of startup option. But in practice even if the argument of
startup option has a '.', check_struct_option() treats it as
key_name. For example if "slow_query_log_file slow1.general_log"
is in configuration file, general log will be turned on.

FIX:

Condition added to make sure that dot is considered as a
part of startup option only if it is before '=' sign and
there is no space before dot.