Fix Bug#16452579 ALTER TABLE ... RENAME INDEX ASSERTS: N == MEM_FIELD_HEADER_GET_LEN(BUF)
This is an UNIV_MEM_DEBUG problem only: we are using the function
mem_heap_get_top() to check if a given string is at the top of a memory
heap. If UNIV_MEM_DEBUG is defined it will assert if the supposed chunk
is not the topmost element. Without the assert the code works correctly
because we check if the returned topmost element is the supposed one.

The solution is to introduce a new function that would just check if a
supposed element is the topmost one in a mem heap.

This patch also contains a new unit test for the newly added function
mem_heap_is_top().

Approved by:	Marko (rb:2135)