BUG#16484323: REPLICATION W/GTIDS FAILS WITH ERROR 1778 WHEN RUNNING ANALYZE OR REPAIR TABLE
After BUG#16062608 fix, when a command that causes a implicit commit
but does not change data is executed when @@SESSION.GTID_NEXT is
different from AUTOMATIC, a error is thrown saying that this command
is not allowed.
Commands ANALYZE and REPAIR TABLE, which cause a implicit commit, were
not correctly identified as update queries, so when them were
replicated slave raised the error.

To fix this incorrect behaviour, ANALYZE and REPAIR TABLE commands are
now identified as update queries.