Followup fix for Bug#16491986 AVOID EXTRA READTUPLE() WHEN UPDATE IS EXECUTED AS DELETE+REINSERT
That fix introduced a bug where all PK columns was not always read 
by the readTuple() prior to the delete. Thus random values was
used to produce the key of the tuple to be deleted - Sometimes
causing the tuple expecting to conflict with our update 
to be deleted instead of the correct tuple.

This fix ensures that 'm_pk_bitmap_p' is always a part
of the read_set whenever delete+reinsert us used
to perform the update.