Bug#16391722-ON 64 BIT WINDOWS SERVER 2012 INNODB IS NOT ALLOCATING              MORE THAN 32GB MEM
This problem was introduced with WL5756 when UNIV_PAGE_SIZE became
mapped to srv_page_size.  Previously, it was a constant number which
gets its type at compile time.  But srv_page_size is a ULONG which is
only 32 bits on Windows 64.  The macro ut_2pow_round(n, m) needs the
same size number in both parameters. When it is used with UNIV_PAGE_SIZE,
it is used with variables of type ulint, which is 64 bits on Windows 64.
On a 64 bit Windows OS, the output would get truncated to 32 bits.

In InnoDB, it is always usefull to typecast UNIV_PAGE_SIZE as a ulint.
When this was done, 5 places were found where UNIV_PAGE_SIZE was used
in a precompiled macro starting with '#if'.  In these cases, the macro
used the declaration value of zero and caused no problems. 

Approved by marko at http://rb.no.oracle.com/rb/r/2182/