Bug #16316828 WL 5092 'REDUCED BINLOG SIZE FOR RBR' CAUSES NDB CONFLICT RESOLUTION FAILURES
For initial releases of mysql-5.6-cluster-7.3, we do not support the
binlog_row_image option - it should have no effect for cluster tables.

We implement this by setting the HTON_NO_BINLOG_ROW_OPT flag which should stop
the replication layer from acting on the binlog row image option.

However, our tests show that this option continues to take effect.

This patch adds a further check in the binlog.cc::binlog_prepare_row_images
function to ensure that no trimming of the write_set occurs when writing
rows to the Binlog, if the handler involved does not have the flag set.

Testing can be done by running with a handler with this flag set, 
varying the binlog row image flag, and checking that the recorded
events are unaffected.  This testing will be pushed separately to 
cluster trees.