Bug#14789787 CRASH WITH UNIQUE BLOBS, DEFAULT VALUE, ON DUPLICATE KEY UPDATE
Problem:-
In an INSERT ... ON DUPLICATE KEY UPDATE statement, using VALUES() function 
with INSERT clause, causes valgrind warnings.

Analysis:-

In an INSERT ... ON DUPLICATE KEY UPDATE statement, we can use the VALUES() 
function in the UPDATE clause to refer to the values that will be 
inserted. In other cases VALUES() return NULL.

Problem occur when we are using VALUES() with the INSERT clause of the INSERT ... ON 
DUPLICATE KEY UPDATE. We are trying to get the referred value, but we 
should give NULL in this case since this VALUES() clause is not part 
of the update clause.

Solution:-
We will find whether this VALUES() is referred in UPDATE clause or INSERT 
clause. According to that, return column value or NULL.