BUG#16502579: READ_GTIDS_FROM_BINLOG IMPROPERLY CALLS MY_ERROR               WITH ER_MASTER_FATAL_ERROR_READ
There were a few problems with how the read_gtids_from_binlog function
returned an error in a very particular and very unlikely scenario. In
detail, the error we are referring to happens when there are
Gtid_log_event(s) in the binlog but on the other hand no
Previous_gtids_log_event are found. This should not happen since
a binary log file with global transaction identifiers should contain
one Previous_gtids_log_event at its beginning and preceding any other
Gtid_log_event in the same file. In fact, if this is not verified, we
should consider it as logical corruption. This invariant is checked
when the server reads the GTIDs from the log files in one go, for
example, when the server restarts, when the dump thread repositions
itself, or when one purges logs.

Now, getting back to the problem at hand... The problems found
regarding how the error was being raised were:

  P1. The wrong ER code was used;
  P2. The wrong error reporting function was used;
  P3. The error was not propagated upwards in the stack;
  P4. PURGE LOGS TO command would silently discard the error when
      forwarding it upwards in the stack.
  P5. In the purge_logs member function, when an error was
      returned by init_gtid_sets, the global_sid_lock was not released.

We fix these problems, by:

  P1 - introducing a new error code. This error code is generic enough
       and refers to logical corruption.
  P2 - using my_printf_error instead of my_error. Thence, we can 
       set the proper arguments to fill in the error format string.
  P3 - propagating the error upwards in the stack, by setting the
       correct return value.
  P4 - introducing error_index and close_error_index in purge_logs to
       differentiate those errors from the error returned by
       init_gtid_sets. Thence we consider all when deciding the return
       value.
  P5 - releasing global_sid_lock whether there was an error or not.