Bug#16182919 MYSQL CLIENT WITH --DEFAULT-CHARACTER-SET=UJIS   CRASHES
MySQL client wrongly used latin1 as charset while
calling wildcmp function for all default client
character sets. The crash occurred when a multibyte
charset is the default client character set, in which
case, the wildcmp implementation calls ismbchar
function. This function, however, is not defined for
latin1 charset and hence the segfault.

#define my_ismbchar(s, a, b)
  ((s)->cset->ismbchar((s), (a), (b)))

Fixed by using the default client charset as the
argument to wilcmp function.