Bug #16497925 - QUERY CACHE BREAKS REPEATABLE READ
Use the read view low limit id transactions because the RO transactions don't
have transaction IDs assigned to them. If a view hasn't been created for a
transaction then it couldn't have done a select. And when a read view is
created for it, its low_limit_id will be the system  max trx id.

Approved by Jimmy Yang rb#2210