BUG#15850951-DUPLICATE ERROR IN REPLICATION WITH SLAVE TRIGGERS AND AUTO INCREMENT
Problem:On filtered slaves, IRU(IntVar, RandVar and UserVar)                    
events are getting executed twice in the span of one query                      
execution time if the query involves in firing a 'AFTER' trigger.                                                                 
                                                                                
Analysis: IRU events are getting cleaned up at the end of                       
the query execution. If a query, with IRU events, involves                      
in firing a 'AFTER' trigger and if the 'AFTER' trigger contains
DML queries, the same IRU events are getting executed once again.                            
This result in unexpected results while inserting tuples in                     
tables which has auto_increment columns.                                        

Fix: Delete these IRU events immediately after it gets                          
exectued for the first time.