Bug #16078466:EXPLAIN CRASH IN FIELD_ITERATOR_TABLE_REF::               GET_OR_CREATE_COLUMN_REF
Analysis:
---------
Server crashes on a debug build when:
a) Update and delete operation is performed using an
invalid column as the join clause within a stored
procedure or function and
b) When the stored procedure or function is invoked a
second time.

In case of SP/SF, the lex object associated with the
instructions are cached when the SP/SF is first invoked.
The flag 'is_join_columns_complete' which is part of the
lex object indicates that the resolution of the join
columns is complete. This flag is set when the SP/SF is
first invoked though the resolution is not complete
since the column in the joining clause does not exist.

Since the lex object is cached, when the SP/SF is invoked
the second time, it is assumed that the resolution is
complete and the server crashes while trying to access
the join column during debug assert.

Note: On a release build, it gives an appropriate error.

Fix:
---
Mark the lex object as broken when there is an error which
forces the lex object to be re-created during the subsequent
invocation of the SP/SF.