Bug#16442113: UNEXPECTED LOG OUTPUT FOR MYSQLD --HELP --VERBOSE
mysqld --help [--verbose] printed various diagnostics to
stderr, which is a fine place for it. When --verbose is
given, we need to go through a full initialization of our
plugins so they can add their variables to the list we'll
print. --help without --verbose only printed one unexpected
line, which we'll now suppress for tidyness. More granularity
for the --verbose case might be gained with the advent of
WL#6661.