Bug #16171537 : IN ELT OF LEAST OF FLOOR OF NULL NULL IS NOT PASSED ON BUT IGNORED, AS IF MIN
Problem:
when we have query like 
select ELT(LEAST(2, FLOOR((NULL + 4) / 7)), 'aaaa','bbbb','cccc','dddd') 
AS Plan;

We get the result

+------+
| Plan |
+------+
| bbbb |
+------+

But in quert like,

select ELT((NULL + 4) / 7, 'aaaa','bbbb');

We get 
+------+
| Plan |
+------+
| NULL |
+------+

Analysis:-

As in case of LEAST(2,NULL), it will set the return value as 2 and with 
that it will set null_value.

But in Item_func_elt::val_str(), when we are checking for 1st argument 
value, we only get the value. We dont check whether this contain a null 
value or not.

Solution:-
After getting the value check whether that expression carry any null value.
If yes, set the current function value as null.