Bug #16522372: START_SLAVE_SQL.INC WAITS FOR THE WRONG THREAD TO START.
start_slave_sql.inc was waiting for the wrong thread to start. It
was waiting for the IO thread when it should be waiting for the
SQL thread.

This patch fixes this.

As a side-effect, we need to fix rpl_grant test, since it was using
start_slave_sql.inc when it should actually use asychronous START
SLAVE SQL_THREAD.