BUG#16517775: MYSQLBINLOG + --INCLUDE-GTIDS + MULTIPLE BINLOG               FILES => UNABLE TO READ
mysqlbinlog is not handling correctly multiple binlog files, when
the GTID switches are used. The problem is that it is skipping
control events like rotate or format description events.

This happens, because, when skipping GTIDs, the flag controlling
the filtering part is only reset on the next GTID event to be
processed. This means that when reaching the end of the file and
thus switching to the next one, ROTATE events and FORMAT
DESCRIPTIONS events are filtered, even though they should
not. Perhaps it would be fine to skip ROTATE events, since they
are output merely as comments, but the FORMAT DESCRIPTION event
is needed for decoding the rest of the file. The best is to not
skip any control event.
            
Furthermore, when using --skip-gtids, Gtid_log_event events are
skipped, and that is OK, but Previous_gtids_log_event events are
not. This does not look consistent.
            
This patch fixes the problem by resetting the filter flag when
processing transaction boundaries events. Control events, such as
START_EVENT_V3, SLAVE_EVENT, STOP_EVENT,
FORMAT_DESCRIPTION_EVENT, ROTATE_EVENT, IGNORABLE_LOG_EVENT,
INCIDENT_EVENT are always processed and printed. In addition,
Previous_gtids_log_event events are now filtered out if
--skip-gtids is used.