Bug#16503490 TABLE-COPYING ALTER TABLE MISTAKES PARTIAL ROLLBACK FOR FULL ROLLBACK
During table-copying online ALTER TABLE (such as ADD COLUMN, DROP
COLUMN, ADD PRIMARY KEY, changing the ROW_FORMAT), we keep a log of
concurrent INSERT and DELETE operations that modify the table data.
Off-page columns (BLOBs) are not logged; instead, we just log the
pointers as they are in the clustered index records.

In order to avoid accessing stale BLOB pointers when applying the log,
we suspend the purge of records that would free BLOBs, and keep track
of user transactions that were rolled back.

The bug is that a partial rollback is wrongly treated as a full
rollback. That is, the remaining changes of a partially rolled back
transaction (one that involved a statement rollback, or an explicit
ROLLBACK TO SAVEPOINT) could be lost (rolled back) during the online
ALTER TABLE.

This patch works around the problem in MySQL 5.6.11. We prohibit
partial rollback during table-copying online ALTER TABLE.

There are two forms of partial rollback.

ROLLBACK TO SAVEPOINT will simply return an error 153, but it will not
commit or roll back the transaction. The user has to explicitly commit
or roll back.

Statement rollback (such as when a duplicate key error is detected
during an INSERT) will fake a deadlock and roll back the entire
transaction.

lock_tables_are_being_altered(): A predicate for checking if a
transaction has locked any tables that are being processed by
table-copying ALTER TABLE. This is similar to lock_table_has(), which
is invoked for the transaction of the currently running thread,
without acquiring any mutex.

innobase_rollback_to_savepoint(): Return HA_ERR_NO_SAVEPOINT if online
ALTER is active.

row_mysql_handle_errors(): Fake a deadlock if online ALTER is active
and we would be rolling back to a savepoint.

We do not take index->lock for the dict_index_is_online_ddl() checks
for performance reasons. The index cannot go away while there is a
transaction holding locks on it.

rb#2233 approved by Kevin Lewis