Bug#16430532 WRONG SIZING HINTS USED IN THE PERFORMANCE SCHEMA
Before this fix, the server did perform the following actions during 
startup.

1) 
Retrieve the value of
- max_connections
- table_definition_cache
- table_open_cache
- open_files_limit
from the user configuration provided in my.cnf,
and use this settings as a base to compute 
performance_schema auto sized variables.

2)
Then, later in the startup process:
- max_connections could be trimmed down
- table_definition_cache default value could be changed
- open_files_limit could be either computed (replacing the 0 value) or 
  adjusted.

Because of the order used, the performance schema automated sizing 
was based on incorrect data.

With this fix, the server startup code has been changed to perform part 2) 
first, and only when all the proper sizing parameters are finally computed 
and adjusted, then perform the performance schema initialization in part 1)

Also, the code performing 2) has been cleaned up to be maintainable.
Clarity in this complicated process (due to related dependencies between 
configuration settings) is critical here.