Fix for Bug#16483273 WRONG RESULT FOR DESCENDING RANGE SCAN USING                      INDEX CONDITION PUSHDOWN
When index condition pushdown was used on a descending range scan
and the first range interval did not contain any qualifying records,
the result of the range scan could be empty even if other 
range intervals contained qualifying records.

When executing a query with query condition like:

  WHERE t1.i1 != 4

as a descending range scan, this will be executed using the
QUICK_SELECT_DESC object. The range optimizer will construct the
ranges "i1 < 4" and "i1 > 4". The QUICK_SELECT_DESC will start with
the last of these ranges and open a descending index scan that
will read all records where i1 > 4. If the query condition has been
pushed to the storage engine using ICP, the code in
QUICK_SELECT_DESC::get_next() will before starting the scan tell the
storage engine that the lower value for this range is "4". This is
done to prevent that the storage engine potentially could read a lot
of records outside of the range interval. 

In the case where this caused wrong result, the initial interval did
not contain any records (or just contained non-qualifying
records). This caused the ICP code inside the storage engine to return
the ICP_OUT_OF_RANGE status, that later is translated to
HA_ERR_END_OF_FILE and is returned to QUICK_SELECT_DESC::get_next() as
the error code from the call to ha_index_last(). This was handled as
if there was no more records in the index, and not as having reached
the end of the specified range. This caused the range scan to stop
after having processed the first range interval.

The fix for this problem is to handle the return code from
ha_index_last() as it is just the end of the current range interval,
not as having reached the end of the index.