Bug #16315767: AES_DECRYPT WITH BAD DATA CAUSES DISCONNECT WHEN USING OPENSSL
When some SSL function fails the code should make sure the SSL error is cleaned up even if
it wants to ignore the actual error.
Because if the error is not cleaned up it will "stick" in some openssl state and will affect later 
SSL calls.
Fixed by making sure aes_encrypt/aes_decrypt are cleaning up after an error. 
Switched to explicit allocation and de-allocation of the cipher context (for clarity what happens
when the openssl function fails).
Test case added.