Bug#16037180 PROBLEM WITH AMOUNT OF SCANNED PAGES FOR FLUSHING IN INNODB
Approved by: Marko Makela and Kevin Lewis rb://2112

When doing a flush list batch the scanning is O(n*n). This is because
once we release buf_pool::mutex we can't trust the pointers when
traversing the flush list. This patch uses the 'hazard pointer'
approach to fix that issue.