Bug #16048546 CRASH: CREATING FEDERATED TABLE WITHOUT OPTIONS
Problem description:
mysqld server is crashed when we start the server with 
federated option, and tries to create the table with out 
connection string.

Analysis:
In get_connection() we are returning wrong error number 
(i.e error_num=1). Hence after processing the error_num, 
we will have error string as 
"Can't create/write to file '%s' (Errcode: %d - %s)". 
The error string is expecting three va_args, but we are 
passing only argument. Which is the cause for the dump.

Fix:
We are setting the correct error number (i.e error_num=1432). 
Hence after processing we will have error string as 
"Can\'t create federated table. The data source connection 
string \'%-.64s\' is not in the correct format". This error 
string is expecting only one va_arg and we are passing that.