Bug#16078792 MY_WRITE() RETURNS WRONG VALUE ON DISK FULL
my_pwrite had the same bug as my_write():

If my_write() hits a disk full condition, it can return a wrong value.

Under certain conditions my_write() tries to repeat to write if it
hits a disk full condition. Eventually it returns "writtenbytes+written"
(if called without MY_NABP | MY_FNABP). 'writtenbytes' is the return
value from the last write(2) call, and 'written' is the sum of bytes
written successfully on earlier iterations.

Now assume, my_write() was called to write 8192 bytes. After 4096 bytes
it hits a disk full condition. That is, the first call to write(2)
returns 4096. 'writtenbytes' is 4096. That value is accumulated to
'written'. In this case my_write() loops to read more data. But now it
gets -1 from write(2) and errno is ENOSPC (for example). my_write()
breaks the loop and returns "writtenbytes+written", which is 4095 now.
(Even though both variables are unsigned, this is the outcome of the
two's complement arithmetics).

4095 is not an appropriate return value in this situation. Since 4096
bytes have been written in fact, that value should be returned to the
caller. The next call to my_write() should then return -1 if the disk
is still full and no single byte could be written.