Fix for Bug#16437630 UNNECESSARY EVALUATION OF QUERY CONDITION WHEN         USING REF ACCESS ON STRING COLUMN
For queries using ref access on CHAR/VARCHAR data types, the ref access
condition was evaluated also as part of the query condition or pushed
down as an index condition to the storage engine. For most data types
this extra evaluation of the ref access condition is unnecessary and
adds overhead.
      
For multiple data types (eg. integers, double) the extra evaluation of
the ref condition was already avoided. This patch extends this to
also avoid the extra evaluation of the ref access condition for CHAR
and VARCHAR fields.