Bug#16522662: SHOW WARNINGS ISN'T RESETTING THE WARNINGS COUNT
The problem was that the first EOF packet sent during execution of
SHOW WARNINGS and SHOW ERRORS contained the number of conditions
raised by the previous statement rather than the number of conditions
raised by SHOW. This bug did not affect the second EOF packet sent
after SHOW completes, only the first EOF packet sent directly after
the initial result set metadata.

This was a regression introduced by WL#6406. This WL changed the
behavior of SHOW WARNINGS | ERRORS so that a new diagnostics area is
pushed on the diagnostics area stack before SHOW executes. This
prevents any conditions raised during SHOW execution from modifying
the diagnostics area we are inspecting, similarly to how
GET DIAGNOSTICS works. The problem was that when the new diagnostics
area is pushed and conditions copied from the old to the new area, the
counter of conditions generated by the current statement was also
updated. And since this new diagnostics area was the current area when
the first EOF packet was sent, the number of conditions generated by
the previous statement was by mistake sent instead of the number of
conditions generated by SHOW.

This patch fixes the problem by resetting the condition counter after
pushing a new diagnostics area for SHOW WARNINGS | ERRORS. A similar
change has been made for GET DIAGNOSTICS (for consistency only, there
were no negative consequences).