WL6742-Improve InnoDB SELECT COUNT(*) performance by using        handler::records()
As a first step toward implementing WL6605 (Improve InnoDB SELECT
COUNT(*) performance), where a known record count is adjusted by
transaction deltas, this worklog implements the handler::record()
call in InnoDB to return the record count to the optimizer instead
of making it count records itself.

This patch improves the cost of an in-memory table scan by about 20%
since record data does not need to be returned to the optimizer via
the handler interface.  It effectively pushes down the counting effort
from the server into InnoDB, reducing a lot of memcpy calls and data
conversions.

Just as before, only records visible to the current transaction are
counted. There is a new testcase in this patch to prove that the
correct count(*) is returned even when other transactions of each
isolation level are in various stages of activity.

Approved by Marko in http://rb.no.oracle.com/rb/r/1693/
QA approval by Amit in http://wl.no.oracle.com/worklog/?tid=6742