Bug#11766777	LIKE WILDCARD MATCH FAILS IN SOME CASES
PROBLEM
--------

The Item_func_like::turboBM_matches() uses Turbo Boyer-Moore  
algorithm which fails to detect some patterns. In the 
traditional brute force methods the search pattern is 
shifted by 1 after each mismatch with the text. Boyer-Moore
uses some rules to determine the amount of shift to the 
search patterns .These rules are

1. Bad character shifts
2. Good character shifts

The maximum shift amount the two is chosen and then used to
shift the seacrh pattern.The Turbo-BM is a modification of
Boyer-Moore algorithm which remembers the last macthed 
search pattern segment before the mismatch occured and
uses this information to do a turbo shift which optimizes 
the shift by skipping the already matched part. The problem
lays in this turbo shift part, for some patterns the shift
is incorrect.The implemantation is according to the 
algorithm but gives wrong results. If we change this logic,
we are not sure how it will effect other search patterns,
since there is no standard data for testing this algorithm.
Therefore we decided to go for much safer Boyer-Moore
algorithm. 
  
FIX
---
We changed the algorithm to implement the original 
Boyer-Moore algorithm instead of Turbo Boyer-Moore 
algorithm.