BUG#13727933- PURGE BINARY LOGS TO MIGHT BE INEFFECTIVE AND NOT VERBOSE ABOUT THAT
Problem:
PURGE BINARY LOGS TO/BEFORE could fail if a binary
log that is expected to be purged is in use or active.
The behavior is fine except that it does not report
this failure and behaves like a query that executed
fine and did purge all the files expected.

Fix:
Fixed by printing a warning if PURGE LOGS fails to delete
a file that is in use or active. Made sure the same warning
is not thrown if it is an automatic purge.