Bug#16528595: MYSQL-TRUNK DOES NOT BUILD ON VISUAL STUDIO 2012 64-BIT
The problem was that trunk compilation on Visual Studio 2012 failed with
  error C3861: 'GetCurrentProcessorNumber': identifier not found

The reason was that GetCurrentProcessorNumber() is only declared
in the Visual Studio 2012 header files if _WIN32_WINNT is >= 0x0502.
This means "Windows Server 2003 with SP1, Windows XP with SP2"
or newer.

The CHECK_FUNCTION_EXISTS() CMake function which was used to check
for the existence of GetCurrentProcessorNumber() seems to use the value
of _WIN32_WINNT matching the platform where we compile. But when
MySQL itself is compiled, we used 0x0501 as defined by Windows.cmake.
This meant that typically the CMake check would pass, but then the MySQL
compilation would fail.

This patch fixes the problem by replacing CHECK_FUNCTION_EXISTS()
with CHECK_CXX_SOURCE_COMPILES() which takes CMAKE_REQUIRED_FLAGS
into account. Then CMAKE_REQUIRED_FLAGS is set to match the value
of _WIN32_WINNT which is used when compiling MySQL. This means that
the same _WIN32_WINNT value is used for the CMake check as when 
compiling MySQL. This value is also changed from 0x0501 to 0x0502 so
that GetCurrentProcessorNumber() can in fact be used. Finally, the patch
moves the Windows specific check to Windows.cmake.

This means that compilation of MySQL now requires at least
"Windows Server 2003 with SP1, Windows XP with SP2" (0x0502)
rather than just
"Windows Server 2003, Windows XP" (0x0501).
Note that neither Windows Server 2003 nor Windows XP is supported
for MySQL starting from 5.6.