Bug #16609256 - INLINE FFLUSH CALL CAUSES DROP IN SYSBENCH PERFORMANCE
The patch for 11763660 expanded the macro for UT_DBG_PANIC which was
defined as abort(). This macro is part of every call to ut_a(), ut_ad()
and uterror.  So this expanded about 7500 lines of code.  The result
was a performance drop because of a generally larger codebase.

So this patch replaces UT_DBG_PANIC with a function called ut_panic().
This patch flushes both stderr and stdout as a precaution before calling
abort.  The previous patch added fflush(stderr) to make sure that all
error messages were written before crashing.

Approved by Marko in http://rb.no.oracle.com/rb/r/2290/