Bug #13864548: MYSQLD_SAFE RESTARTS MYSQLD AFTER A CRASH WHEN   /ETC/INIT.D/MYSQL STOP IS
This bug is about finding a way to reliably stop mysqld_safe.sh script from 
re-starting the server.
This is achieved by "/etc/init.d/mysql stop" making an empty "flag" file on a
predefined location that, when present, would cause mysqld_safe.sh to
exit without restarting the server.
 
The /etc/init.d/mysql stop script enchanced to create an empty file next to 
the pid file : $pidfile.shutdown
The mysqld_safe script enhanced to look for this empty file, and if present,
delete it and exit instead of restarting the server.
Delete the .shutdown file at the end of mysqld_safe.sh