Bug #16488043:  set password and set password for .. behaves differently for the same user..
<username>@'' is a special edge case of a host name equivalent
in functionality to <username>@'%'.
SET PASSWORD = reuses the SET PASSWORD FOR ... = code
through a special case of taking the user name in the parser 
and keeping the host name empty (NULL), thus signaling the 
execution code that it needs to fill in the current user's host
name.
But in doing this the execution code incorrectly assumes that 
empty string host name of the current user must be expanded to
'%' before the comparison.
Thus the subsequent check in the SET PASSWORD FOR ... = 
code failed to identify the inferred user as the logged in user and
threw an error when the user doesn't have enough privileges
to write to the mysql.user table (basically what unrestricted 
SET PASSWORD FOR ... = needs).

Fixed by not substituting an empty logged in user host name 
with a '%' in this case.

Test case added.